Updating from master

+++ neutron/agent/linux/daemon.py
@@ -15,16 +15,31 @@
 import atexit
 import fcntl
 import grp
+import logging as std_logging
+from logging import handlers
 import os
 import pwd
 import signal
 import sys
 
+from oslo_log import log as logging
+
 from neutron.common import exceptions
 from neutron.i18n import _LE, _LI
-from neutron.openstack.common import log as logging
 
 LOG = logging.getLogger(__name__)
+
+DEVNULL = object()
+
+# Note: We can't use sys.std*.fileno() here.  sys.std* objects may be
+# random file-like objects that may not match the true system std* fds
+# - and indeed may not even have a file descriptor at all (eg: test
+# fixtures that monkey patch fixtures.StringStream onto sys.stdout).
+# Below we always want the _real_ well-known 0,1,2 Unix fds during
+# os.dup2 manipulation.
+STDIN_FILENO = 0
+STDOUT_FILENO = 1
+STDERR_FILENO = 2
 
 
 def setuid(user_id_or_name):
@@ -53,6 +68,25 @@ def setgid(group_id_or_name):
             msg = _('Failed to set gid %s') % new_gid
             LOG.critical(msg)
             raise exceptions.FailToDropPrivilegesExit(msg)
+
+
+def unwatch_log():
+    """Replace WatchedFileHandler handlers by FileHandler ones.
+
+    Neutron logging uses WatchedFileHandler handlers but they do not
+    support privileges drop, this method replaces them by FileHandler
+    handlers supporting privileges drop.
+    """
+    log_root = logging.getLogger(None).logger
+    to_replace = [h for h in log_root.handlers
+                  if isinstance(h, handlers.WatchedFileHandler)]
+    for handler in to_replace:
+        new_handler = std_logging.FileHandler(handler.baseFilename,
+                                              mode=handler.mode,
+                                              encoding=handler.encoding,
+                                              delay=handler.delay)
+        log_root.removeHandler(handler)
+        log_root.addHandler(new_handler)
 
 
 def drop_privileges(user=None, group=None):
@@ -97,8 +131,7 @@ class Pidfile(object):
         return self.pidfile
 
     def unlock(self):
-        if not not fcntl.flock(self.fd, fcntl.LOCK_UN):
-            raise IOError(_('Unable to unlock pid file'))
+        fcntl.flock(self.fd, fcntl.LOCK_UN)
 
     def write(self, pid):
         os.ftruncate(self.fd, 0)
@@ -133,28 +166,41 @@ class Daemon(object):
 
     Usage: subclass the Daemon class and override the run() method
     """
-    def __init__(self, pidfile, stdin='/dev/null', stdout='/dev/null',
-                 stderr='/dev/null', procname='python', uuid=None,
-                 user=None, group=None):
+    def __init__(self, pidfile, stdin=DEVNULL, stdout=DEVNULL,
+                 stderr=DEVNULL, procname='python', uuid=None,
+                 user=None, group=None, watch_log=True):
+        """Note: pidfile may be None."""
         self.stdin = stdin
         self.stdout = stdout
         self.stderr = stderr
         self.procname = procname
-        self.pidfile = Pidfile(pidfile, procname, uuid)
+        self.pidfile = (Pidfile(pidfile, procname, uuid)
+                        if pidfile is not None else None)
         self.user = user
         self.group = group
+        self.watch_log = watch_log
 
     def _fork(self):
         try:
             pid = os.fork()
             if pid > 0:
-                sys.exit(0)
+                os._exit(0)
         except OSError:
             LOG.exception(_LE('Fork failed'))
             sys.exit(1)
 
     def daemonize(self):
         """Daemonize process by doing Stevens double fork."""
+
+        # flush any buffered data before fork/dup2.
+        if self.stdout is not DEVNULL:
+            self.stdout.flush()
+        if self.stderr is not DEVNULL:
+            self.stderr.flush()
+        # sys.std* may not match STD{OUT,ERR}_FILENO.  Tough.
+        for f in (sys.stdout, sys.stderr):
+            f.flush()
+
         # fork first time
         self._fork()
 
@@ -167,22 +213,23 @@ class Daemon(object):
         self._fork()
 
         # redirect standard file descriptors
-        sys.stdout.flush()
-        sys.stderr.flush()
-        stdin = open(self.stdin, 'r')
-        stdout = open(self.stdout, 'a+')
-        stderr = open(self.stderr, 'a+', 0)
-        os.dup2(stdin.fileno(), sys.stdin.fileno())
-        os.dup2(stdout.fileno(), sys.stdout.fileno())
-        os.dup2(stderr.fileno(), sys.stderr.fileno())
+        with open(os.devnull, 'w+') as devnull:
+            stdin = devnull if self.stdin is DEVNULL else self.stdin
+            stdout = devnull if self.stdout is DEVNULL else self.stdout
+            stderr = devnull if self.stderr is DEVNULL else self.stderr
+            os.dup2(stdin.fileno(), STDIN_FILENO)
+            os.dup2(stdout.fileno(), STDOUT_FILENO)
+            os.dup2(stderr.fileno(), STDERR_FILENO)
 
-        # write pidfile
-        atexit.register(self.delete_pid)
-        signal.signal(signal.SIGTERM, self.handle_sigterm)
-        self.pidfile.write(os.getpid())
+        if self.pidfile is not None:
+            # write pidfile
+            atexit.register(self.delete_pid)
+            signal.signal(signal.SIGTERM, self.handle_sigterm)
+            self.pidfile.write(os.getpid())
 
     def delete_pid(self):
-        os.remove(str(self.pidfile))
+        if self.pidfile is not None:
+            os.remove(str(self.pidfile))
 
     def handle_sigterm(self, signum, frame):
         sys.exit(0)
@@ -190,7 +237,7 @@ class Daemon(object):
     def start(self):
         """Start the daemon."""
 
-        if self.pidfile.is_running():
+        if self.pidfile is not None and self.pidfile.is_running():
             self.pidfile.unlock()
             LOG.error(_LE('Pidfile %s already exist. Daemon already '
                           'running?'), self.pidfile)
@@ -205,4 +252,6 @@ class Daemon(object):
 
         start() will call this method after the process has daemonized.
         """
+        if not self.watch_log:
+            unwatch_log()
         drop_privileges(self.user, self.group)
